+++ web/components/layout_templ.go
@@ -55,7 +55,7 @@ func Layout(title string, description string) templ.Component {
 		if templ_7745c5c3_Err != nil {
 			return templ_7745c5c3_Err
 		}
-		templ_7745c5c3_Err = templruntime.WriteString(templ_7745c5c3_Buffer, 3, "\"><link rel=\"stylesheet\" href=\"/static/css/app.css\"></head><body class=\"min-h-screen bg-background font-sans text-foreground antialiased\"><a class=\"skip-link\" href=\"#conteudo\">Pular para o conteúdo</a>")
+		templ_7745c5c3_Err = templruntime.WriteString(templ_7745c5c3_Buffer, 3, "\"><link rel=\"icon\" type=\"image/png\" href=\"/static/images/branding/logo-printlab-primary.png\"><link rel=\"stylesheet\" href=\"/static/css/app.css\"></head><body class=\"min-h-screen bg-background font-sans text-foreground antialiased\"><a class=\"skip-link\" href=\"#conteudo\">Pular para o conteúdo</a>")
 		if templ_7745c5c3_Err != nil {
 			return templ_7745c5c3_Err
 		}